backend/requirements.txt
flask==2.3.3
Werkzeug==2.2.2
elasticsearch==7.17.9
pyyaml==6.0
python-dotenv==1.0.0
requests==2.28.2
flask-cors==3.0.10
grpcio==1.59.0
grpcio-tools==1.59.0
protobuf==4.24.4 